Create Count S1 in S2
The program must accept two string values S1 and S2 as the input where the length of S2 is always 2. 
The program print the number of occurrences of the string S2 in S1 as the output. 
Boundary Condition(s): 2 <= Length of S1 <= 100 
Example Input/Output 1: Input: tomato to 
Output: 2
 Example Input/Output 2: Input: AABBBCCBB BB 
Output: 3