Create Class Juice
The program accepts the name and price of a juice as the input. The program prints the name of the juice and prints the price of 5 juices as the output. Please fill in blanks with code so that the program runs successfully. 
Example Input/Output 1: Input: Orange 35 Output: Orange 175 Example Input/Output 2: Input: Apple 80 Output: Apple 400