Create Maximum Element In The Matrix
Input Format:
The input consists of (m*n+2) integers. The first integer corresponds to m, the number of rows in the matrix and the second integer corresponds to n, the number of columns in the matrix. The remaining integers correspond to the elements in the matrix. The elements are read in row wise order, first row first, then second row and so on. Assume that the maximum value of m and n is 10.
Output Format:
Refer sample output for details.
Sample Input 1: 
3
2
4 5
6 9
0 3
Sample Output 1:
The maximum element is 9